Use reference counting to manage open file handles.
Make file handle close action predictable and avoid having file open for long time. Relying on GC is not flexible and do not guarantee that file are closed after DB was closed. We do not even implement "finalizer" method as external iterators are already required to be closed by end user and this method is deprecated in Java 9.