fix: compact all boundary files that share same user key
As reported in Github issue 339 in google/leveldb, it is incorrect to split the
same user key across multiple compacted files since it causes
tombstones/newer-versions to be dropped, thereby exposing obsolete
data. There was a fix by commit 9ffeed11, but it ended up not fully fixing
the problem. (It checked for boundary problems in the first level
being compacted, but not the second). This problem was revealed
by Github issue 887 in google/leveldb.

We now adjust boundaries to avoid splitting user keys in both the
first level and the second level.